basket half done

+++ src/main/java/introinheritance/employee/Employee.java
@@ -13,6 +13,6 @@ public class Employee extends Person{
     }
 
     public void raiseSalary(int percent){
-        salary = salary + (salary * (percent/100));
+        salary = salary +  (salary * ((double) percent/100));
     }
 }
